requirements.txt
# to build apk install briefcase

toga-android~=0.4.2
requests~=2.31.0
